Fix createURL
The Javascript URL object takes a second argument which is the base to
use for the URL. However it only accepts the "root" of the base. So
arguments such as http://localhost/example get silently turned into
http://localhost/

This change places the baseURL within the path element, after making
sure that it ends with a slash (/).