Minor (subjective) fixup to Ukrainian translation
There is 2 words with last letter "Ы" which does not exist in Ukrainian alphabet and therefore these words are grammatically incorrect. Changed all two instances with "и". 
The words "Upload" and "Download" in this translation are translated as the same word ("Завантажити") which is theoretically a right translation for both words, but may confuse a user for a second(it did me at least). Proposed solution here is to translate "Upload" as "Вивантажити".